Auth validation

+++ app/Http/Controllers/MusicaController.php
@@ -6,6 +6,7 @@ use App\Musica;
 use App\File;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Input;
 
 class MusicaController extends Controller
 {
@@ -48,8 +49,41 @@ class MusicaController extends Controller
             'autor' => 'required'
         ]);
 
-        Musica::create($request->all());
-        session()->flash('info', 'Cadastro de música realizado com sucesso!');
+        // /**
+		// * Musica Model Creation
+		// */
+        $musica = new \App\Musica();
+        $musica->nome = $request->nome;
+        $musica->autor = $request->autor;
+        $musica->save();
+
+        // /**
+		// * Retrieving File from form
+		// */
+        $file = Input::file('arquivo'); 
+    
+        if(isset($file)){
+            // /**
+            // * Storage related
+            // */
+            $storagePath = storage_path().DIRECTORY_SEPARATOR.'app'.DIRECTORY_SEPARATOR.'public'.DIRECTORY_SEPARATOR.$musica->id;    
+            $fileName = 'track'. $musica->id.'.mp3';
+
+            
+            // /**
+            // * File Model Creation
+            // */
+            $fileModel = new \App\File();
+            $fileModel->name = $file->getClientOriginalName();
+            $musica->file()->save($fileModel);
+            
+            // /**
+            // * Save File 
+            // */
+            $file->move($storagePath, $fileName);
+        }
+
+        session()->flash('success', 'Cadastro de música realizado com sucesso!');
         return redirect('/admin/musicas');
     }
 
@@ -57,7 +91,7 @@ class MusicaController extends Controller
         // /**
 		// * Request related
 		// */
-		$file = \Request::file('documento');
+		$file = \Request::file('arquivo');
 		$musicId = \Request::get('musicId'); 
         
 		// /**
@@ -84,6 +118,7 @@ class MusicaController extends Controller
             $musica->file()->save($fileModel);
         }
 
+        session()->flash('success', 'Upload realizado com sucesso!');
         return $file->move($storagePath, $fileName);
     }
 
@@ -112,7 +147,7 @@ class MusicaController extends Controller
         $musica->nome = $request->nome;
         $musica->autor = $request->autor;
         $musica->save();
-        session()->flash('info', 'Musica atualizada com sucesso!');
+        session()->flash('success', 'Musica atualizada com sucesso!');
         return redirect('/admin/musicas');
     }
 
@@ -124,22 +159,27 @@ class MusicaController extends Controller
      */
     public function destroy(Musica $musica)
     {
-        $dir = $this->public_storage($musica->id);
-        $file = $dir.DIRECTORY_SEPARATOR."track$musica->id.mp3";
-        // /**
-		// * File delete
-		// */
-        unlink($file);
-        // /**
-		// * Directory delete
-		// */
-        rmdir($dir);
         
-        // /**
-		// * File register delete
-		// */
-        File::destroy($musica->file->id);
+        if(isset($musica->file)){
+
+            $dir = $this->public_storage($musica->id);
+            $file = $dir.DIRECTORY_SEPARATOR."track$musica->id.mp3";
+            // /**
+            // * File delete
+            // */
+            unlink($file);
+            // /**
+            // * Directory delete
+            // */
+            rmdir($dir);
+            
+            // /**
+            // * File register delete
+            // */
+            File::destroy($musica->file->id);
         
+        }
+
         // /**
 		// * Music register delete
 		// */
